Refactored folder drag-and-drop and pagination logic to improve encapsulation and simplify main pager components:
- Created `FolderDragGridItemHelper.kt` to isolate folder-specific drag handling logic, moving it out of the general `DragGridItemHelper.kt`.
- Updated `FolderScreen` to manage its own `PagerState` using `rememberPagerState` and handle local drag events via a new `LaunchedEffect`.
- Simplified `PagerScreenState.handleDragGridItem` and `DragGridItemHelper.handleDragGridItem` signatures by removing numerous folder-related parameters.
- Relocated the `handlePageDirection` helper function from `DragGridItemHelper.kt` to `HomeUtil.kt` for better visibility and reusability.
- Moved folder dismissal logic triggered by scroll progress directly into the `FolderScreen` component.
- Cleaned up unused imports and standardized formatting in `PagerScreen.kt` and `DragGridItemHelper.kt`.
